+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- feat: add `--configure` option
+
 ## [v2.1.6] - 2025-01-05
 
 - refactor: options building